pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mvp-agent"
version = "0.1.0"
description = "Local-first financial news and macro literature-review assistant."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML==6.0.2",
  "pypdf>=5,<6",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10,<2",
  "ruff>=0.9,<1",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
files = ["apps"]
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.setuptools.packages.find]
include = ["apps*"]
exclude = ["artifacts*", "evals*", "tests*"]
namespaces = true
